workaround upscaled blur pixelation (#97)
the old way used in menu blur is kinda hacky (cuz i found the games's original blur is
linear sampled 5x5, not jump lines), and seems to have some negative effect when used on bloom (pixelation)
so back to increasing taps, however as the image size goes up, it becomes quite harder/impossible to get the same blurry result for some reason. 5x5 menu blur ends up in 100 taps in 10k to get the similar blur, and 9x9 used in bloom is even blurrier... so i give up trying to get the same blur by increasing a lot taps, but just scale linearly to fix pixelation first and wait for a better solution.
also the texturesize() returns weird value when mipmap is present, so changed to calculate the image size from the data passed from vertex shader (as we already figured out what blur is used so offset is known).