Splitted SM3DW and CTTT. Added partial DoF and Bloom fix.
Splitted Super Mario 3D World and Captain Toad: Treasure Tracker as both are getting their own shaders. Both are partial due to mips which are hard to scale and due to the rarity of some of the bloom shaders (only in the latest world. Might fix them later but they're awfully long and weird)