Shortened and Colorized Appveyor output (#120)
* Console output compacted and colorized
The installation info took too many lines and won't be revamped while working great as is, so I'd think that it's better to supress any logging. The quiet-mode of the installation doesn't really seem to work either.

The installation part that sometimes errors is irrelevant to the part I've suppressed.