Website Download Improvements
Should feel smoother due to the downloads being downloaded in the tab itself. Scrolls to credit when downloaded. Has a slightly reduced amount of junk code, but only slightly so don't worry 'bout it. Also faster execution if you're visiting the site normally (so if you didn't use the #directDownload hash and view the website) due to the download url fetch only being done when you either click the button or have the specific download hash.