Add ClarityGFX pages via submodule, update some text.
The clarity pages work via submodules which should be controlled via MelonSpeedruns branch. The url will still belong to this repository. One minor point is that to retrieve his latest updates, you need to update the submodule via a console with the command `git submodule update --remote --merge` in the github branch. I'll try to take care of this whenever there's an update.

Due to the _data setup it's really easy to add Presets (https://github.com/MelonSpeedruns/ClarityGFX_Presets/blob/master/_data/presets.yml). The photo's can be clicked on to get a magnified look at it. Also, it doesn't have to be game exclusive. The website might be expended with more features, however I don't know what to add right now. The content will be replaced by real presets, it's all placeholders for now.