Reverted back some old behaviour, improvements!
Not sure if this works as my fork currently has a pull request. Should work. The old behaviour was to fetch the call regardless if people click the link. Upside is that the link will now store the version number of the last date you visited the website if you don't have any 'do not track' setting turned on (which generally isn't the case). Then it'll check and change the text a bit with the current version. Also has some small automatisation improvements for people that have a bot. 

The CSS is also changed so that the download buttons are more prominent, as requested hehe.