[XCX] Alignment cleanup + 8k fix inital commit
Fixes alignment drifting on higher resolutions.  By replacing with fixes
earlier in the rendering pipe.
This should also improve a lot of light leaking, but full res tweak will
be in upcoming commits.

As this is the initial commit, skell view may still be off and some
specific shadow fixes will probably be retired.