Rename Fluid to Glitch Artifacts & add ice artifacts fix
This fix might only apply to artifacts found on the reflective ice on a frost talus fight.
This pack also got renamed due to being used in many more spots besides lava and water. I don't have a great name for it, so I choose this because it's easy to understand for normal people without a clue.