Some Updates to Super Smash Brothers (#265)
* Update 315d61ad21f97614_0000000000000000_vs.txt

Proper Cast the variables  as floats because they were failing to compile.

* Update 4b92e636153d6b54_0000000000000000_vs.txt

Proper Cast the variables as floats because they were failing to compile.

* Delete 1ed6dd85057f3924_0000000000000000_vs.txt

It is not needed as editing this shader won't fix the issue.

* Add files via upload

Fragment Shader that when modified does the same thing as resizing the pixel mess texture. However the shader is reloadable.

* Update rules.txt

Presets to load hide the pixel mess broken effect and set it back to original behavior.