Super Smash Bros Pixel-Mess Revert (#266)
* Delete 42f0844ab0996fe2_0000000000000079_ps.txt

This shader allows for changing of the color for the pixel-mess texture and or making it transparent/opaque. However its also responsible for the Cursor Text in selection screen. So messing with this Shader also messes with the Cursor text. The best way to hide the pixel-mess is to do it the old "version 2" way.

* Update rules.txt

The best way to hide the pixel-mess is to do it the old "version 2" way. Messing with shaders also affects the cursor in selection screen.