[BotW] Switch to Xalphenos' FPS++ V2.1 (fix arrow time)
Switch over to Xalphenos version since his version now has a dynamically sized "oldest frame" (there's not really a good name for what it does) thing which was one of the main things in my version.

Once FPS++ progress is halted, I'll clean up the comments again to something similar what I had in my version, which I feel is more understandable for people that want to see what it does and maybe improve it.

If testing is correct, the previous version also had stability drops or something in comparison with this version. If anything, this update only pushes the arrow time stamina drain fix.