[XCX] Switch exposure from AMD and Nvidia
They were switched from the previous versions. Besides that, somebody reported it being different then before, but I can't find the difference. (There's a slight difference because the vibrancy in one shader was 0.1375 instead of 0.138 but I doubt this would be a big difference and it does simplify it a little bit).