[TPHD*] Fix menu background
Accidentally applied a suggested fix to the wrong game. Reverted that and applied it to the correct game. Also changed the description of the performance pack slightly so that people know what it does. Could probably be more conclusive about what it does. Looks like a regular shadow resolution pack to me.