[DKC:TF] Fix missing puzzle pieces
Previously wouldn't show puzzle pieces in the level selection screen.

The texture that would also need to be scaled is 0x033 which is a compressed texture which can't be scaled, so removing this filter was the only option. But this was mostly pointless anyway since the GUI is also just images which won't really become sharper.

Issue reported by Nathan in the discord.