Merged Workaround Packs
I merged the workaround packs for BOTW into GPU catagories for OpenGL and updated their version to 4.
Along with this removed the workarounds for intel flipped lights and reduced bloom as I tested and concluded that its a non issue on latest drivers for iGPUs
I also updated LWZX Crash pack to V4 as well.