[MK8] Improve descriptions, fix folder spelling, simplify presets
Since filters don't work with Vulkan (even if you try to adapt it to Vulkan) it would be smart if we made the pack OpenGL only at least, since Vulkan will crash with it if you've got it enabled. We want to keep the game breaking to a minimum.

I also changed the descriptions to be more explanatory, and changed the preset names to make more sense to users. Unsure what the presets exactly entail (more specifically the preset that was originally called Smooth half original).