[BotW] Removes 2 old Intel GPU workarounds (#434)
These workarounds were intended for fixing the issues caused by the old WDDM 23 drivers, which offered better performance than their newer counterparts and was able to boot in game prior to Cemu 1.15.17. Now, since BotW is able to boot in game with the newer drivers with similar or better performance, these 2 can be removed as the issues were fixed in WDDM 24 and newer drivers. Intel even removed the DL link to older drivers like the WDDM 23 and 24 ones.