More Splatoon Patches
The "Better Squid Sisters Lighting" patch just removes the normal news lighting so it's set to what it would normally look like in that position in the plaza, making it look "somewhat" better. This also includes a patch the removal of the 2D news layer, and a special swapper, all made by me.