[BotW] More proper faster arrow fix
Since 1.0 floats are very common, this float was likely used in other game code as well. Changing it to 3.0 could be bad, so it's better to just change the instruction that loads it and does the arrow loading.
The fix here is to just have a symbol available to load the lower half from, instead of hardcoding it which Cemu's backwards compatible patcher doesn't like and softlocks.