[Smash 4] Revert border fix
Apparently this border fix is exploiting some undefined behavior or something because it crashes, misrenders GUIs or just shows black on some APIs. Just reverted it to the previous state, except that I kept the resolutions in.

Hopefully a better border fix, or potentially just a workaround, can be found.