[Child Of Light] Fix inventory screen using new Cemu 1.22.5d features
Fixes https://github.com/ActualMandM/cemu_graphic_packs/issues/501

Basically, the game uses the 512x512 for both inventory items and for the main character's rendering. So I told Exzap who added this inMEM1 variable which is perfect for cases like these. Anyway, this is basically just the implementation of that.