E fix: Online Griefing Patch
It was discovered that the patch_offline_squad could be used to destroy currently active online squad missions that other players are doing. This Fixes that by making the mod kill the online. This also provides some troubleshooting text in the rules.txt if the mod doesn't work for a player.

Ideally a better kill online code should exist but this is the best one that exist rn