[Xenoblade X] Adds support for unused weather conditions | Also fixes a mod that could be used for online griefing (#580)
* Update rules.txt

Adds support for unused weather conditions in the game

* Update patch_exp_blade.asm

fixes 1.0.2J support

* Update rules.txt

Felt I was credited enough in the github and blingblong did most of the work, I just got the mod in a state where it was polished enough to be published

* Update rules.txt

in v886 I made TA the default instead of DLC to fix a bug it caused in the JP version. This is making TA first in the name to match TA being the first option

* Update rules.txt

* Add files via upload

Small fix with previous commit. this completes everything I wanted to do

* final weather improvement

Actual final commit. Made it so that the preset names now show the $wrt value making easier to understand, and added support for the blade barracks weather conditions (just need to uncomment to use)

* E fix: Online Griefing Patch

It was discovered that the patch_offline_squad could be used to destroy currently active online squad missions that other players are doing. This Fixes that by making the mod kill the online. This also provides some troubleshooting text in the rules.txt if the mod doesn't work for a player.

Ideally a better kill online code should exist but this is the best one that exist rn

* Update rules.txt

Typo fix... oops

* globalnemesis offline feature update

you can now pick between the friday version and the monday version of global nemesis telethia plume. (previously the mod only had monday, which is worse than friday. now it has both)