update some XCX mods and add some new mods
adding some new mods. sorry the character arts mod rules.txt is 11000 lines.
it needs to be that way to work properly and it does currently work properly
if I had better formatting I may be able to get it down to 9000 lines but not worth it
also the conditional statements are so gnarly because I added support to change the mod behavior is another mod comes allong to either get a different weapon on the character or find a way arround the game preventing characters unlocking arts they shouldn't have.
all it takes to activate that support is uncomment 4 pressets

I was working on a similar mod for skills last year. knew what I was doing that time so formatted it cleanly and skills are less complicated than arts so it was 5000 lines
but my harddrive died and I lost my motivation to return to it or graphics packs in general

also hid the native aa option in the antialiasing pack
that pack was all kinds of broken, aa removal didn't work, fixed that
fxaa wasn't working fixed that
tried to get native aa mod to not produce massive ammounts of artifacts but couldn't