[Splatoon] Add "Disable Autorotate" Mod (#697)
This removes the function call that rotates the camera when the player moves left or right. This makes it much easier to control your aim with both motion and mouse (via Xapfish) since you don't have to fight against the camera when you move.

Thanks for @Gloomy31 adding this pack!