[Skylanders] Add the "Enable Debug" in tfbTool games
For Spyro's Adventure, Giants and Trap Team.

Note: Trap Team does not have a debug folder with the files leftover in the retail game, so most of the debug features are missing (there are some still in some places, like being able to change the hub state in the Academy).
We do have the debug files from Trap Team thanks to the mobile versions but they're not compatible. If the modding tools lets us port them to the Wii U version one day, I'll add them to the pack.